+++ lib/index.js
@@ -1,0 +1,66 @@
+/**!
+ * urlencode - lib/urlencode.js
+ *
+ * Copyright(c) 2012 - 2014
+ * MIT Licensed
+ *
+ * Authors:
+ *   fengmk2 <fengmk2@gmail.com> (http://fengmk2.github.com)
+ */
+
+"use strict";
+
+/**
+ * Module dependencies.
+ */
+
+var iconv = require('iconv-lite');
+
+function isUTF8(charset) {
+  if (!charset) {
+    return true;
+  }
+  charset = charset.toLowerCase();
+  return charset === 'utf8' || charset === 'utf-8';
+}
+
+function encode(str, charset) {
+  if (isUTF8(charset)) {
+    return encodeURIComponent(str);
+  }
+
+  var buf = iconv.encode(str, charset);
+  var encodeStr = '';
+  for (var i = 0; i < buf.length; i++) {
+    encodeStr += '%' + buf[i].toString('16');
+  }
+  encodeStr = encodeStr.toUpperCase();
+  return encodeStr;
+}
+
+function decode(str, charset) {
+  if (isUTF8(charset)) {
+    return decodeURIComponent(str);
+  }
+
+  var bytes = [];
+  for (var i = 0; i < str.length; ) {
+    if (str[i] === '%') {
+      i++;
+      bytes.push(parseInt(str.substring(i, i + 2), 16));
+      i += 2;
+    } else {
+      bytes.push(str.charCodeAt(i));
+      i++;
+    }
+  }
+  var buf = new Buffer(bytes);
+  return iconv.decode(buf, charset);
+}
+
+
+module.exports = encode;
+module.exports.encode = encode;
+module.exports.decode = decode;
+module.exports.parse = require('./parse');
+module.exports.stringify = require('./stringify');
