+++ lib/parse.js
@@ -1,0 +1,84 @@
+var utility = require('utility');
+var decode = require('./index').decode;
+
+function parse(qs, sep, eq, options) {
+  if (typeof sep === 'object') {
+    // parse(qs, options)
+    options = sep;
+    sep = null;
+  }
+
+  sep = sep || '&';
+  eq = eq || '=';
+  var obj = {};
+
+  if (typeof qs !== 'string' || qs.length === 0) {
+    return obj;
+  }
+
+  var regexp = /\+/g;
+  qs = qs.split(sep);
+
+  var maxKeys = 1000;
+  var charset = null;
+  if (options) {
+    if (typeof options.maxKeys === 'number') {
+      maxKeys = options.maxKeys;
+    }
+    if (typeof options.charset === 'string') {
+      charset = options.charset;
+    }
+  }
+
+  var len = qs.length;
+  // maxKeys <= 0 means that we should not limit keys count
+  if (maxKeys > 0 && len > maxKeys) {
+    len = maxKeys;
+  }
+
+  for (var i = 0; i < len; ++i) {
+    var x = qs[i].replace(regexp, '%20');
+    var idx = x.indexOf(eq);
+    var kstr, vstr, k, v;
+
+    if (idx >= 0) {
+      kstr = x.substr(0, idx);
+      vstr = x.substr(idx + 1);
+    } else {
+      kstr = x;
+      vstr = '';
+    }
+
+    if (kstr && kstr.indexOf('%') >= 0) {
+      try {
+        k = decode(kstr, charset);
+      } catch (e) {
+        k = kstr;
+      }
+    } else {
+      k = kstr;
+    }
+
+    if (vstr && vstr.indexOf('%') >= 0) {
+      try {
+        v = decode(vstr, charset);
+      } catch (e) {
+        v = vstr;
+      }
+    } else {
+      v = vstr;
+    }
+
+    if (!utility.hasOwnProperty(obj, k)) {
+      obj[k] = v;
+    } else if (Array.isArray(obj[k])) {
+      obj[k].push(v);
+    } else {
+      obj[k] = [obj[k], v];
+    }
+  }
+
+  return obj;
+}
+
+module.exports = parse;
